refactor(data-table): move sub-components into data-table/ folder
